refactor: update color constants in avatar component, improve naming dashboard skeleton and use chpllinks
[#OCD-4956]
