+++ src/app/components/util/index.js
@@ -1,7 +1,7 @@
 import ChplAvatar from './avatar';
 import ChplConfirmation from './confirmation';
 import ChplCriterionTitle from './criterion-title';
-import ChplDashboardSkeleton from './chpl-dashboard-skeleton';
+import ChplDashboardSkeleton from './dashboard-skeleton';
 import ChplDialogTitle from './chpl-dialog-title';
 import ChplEllipsis from './chpl-ellipsis';
 import ChplLink from './chpl-link';
